Fix: move identity constants to fixtures
